requirements.txt
pillow
opencv-python
gTTS
playsound
dlib
face_recognition